--- a/python/performance/point_update_performance.py
+++ b/python/performance/point_update_performance.py
@@ -1,6 +1,14 @@
+import sys
+from os import pardir
+from os.path import dirname, abspath, join
+# we need to add the parent directory to the sys.path in order to import from
+# other modules
+sys.path.append(join(dirname(__file__), pardir))
 from dataclasses import dataclass, replace
 from typing import List, Tuple
 from time import time
+from copy import deepcopy
+from utils.timeit import timeit
 import random
 
 
@@ -15,9 +23,11 @@ The options are:
 2. Use a class, and mutate the instance for each update.
 3. Use a class, and construct a new instance for each update.
 4. Use a dataclass, and use `replace` to construct a new instance for each
-update
+   update
 5. Use a dataclass, and manually construct a new instance for each update
 6. Use a dataclass, and mutate the instance for each update
+7. Use a dataclass, and deepcopy the instance, and then mutate it for each
+   update
 
 
 The performance of each of these options (in seconds, so large numbers are
@@ -29,11 +39,12 @@ bad):
 4. profile_point_dataclass0: 1.08
 5. profile_point_dataclass1: 0.36
 6. profile_point_dataclass2: 0.18
+7. profile_point_deepcopy: 6.36
 
 Given the cleaner syntax of dataclasses, and equivalent performance, we
-recommend using dataclasses wherever possible. However, whether to mutate them
-or construct new instances is up to you (if the tradeoff of using mutability is
-worth the ~2x performance gain).
+recommend using dataclasses wherever possible. Whenever possible, construct
+new instances rather than mutating in place, in order to avoid deepcopy()ing
+later.
 """
 
 
@@ -86,6 +97,14 @@ def profile_point_dataclass2(initial: Tuple[int, int], values: List[Tuple[int, i
         point.x = x0
         point.y = y0
 
+def profile_point_deepcopy(initial: Tuple[int, int], values: List[Tuple[int, int]]):
+    point = DataPoint(initial[0], initial[1])
+    for (x0, y0) in values:
+        copy = deepcopy(point)
+        copy.x = x0
+        copy.y = y0
+        point = copy
+
 
 if __name__ == "__main__":
     xs = random.sample(range(round(1e6)), round(1e6))
@@ -93,27 +112,10 @@ if __name__ == "__main__":
     vals = list(zip(xs, ys))
     init = vals[0]
     rest = vals[1:]
-    t = time()
-    profile_tuple(init, rest)
-    delta = time() - t
-    print("profile_tuple: {}".format(delta))
-    t = time()
-    profile_point0(init, rest)
-    delta = time() - t
-    print("profile_point0: {}".format(delta))
-    t = time()
-    profile_point1(init, rest)
-    delta = time() - t
-    print("profile_point1: {}".format(delta))
-    t = time()
-    profile_point_dataclass0(init, rest)
-    delta = time() - t
-    print("profile_point_class0: {}".format(delta))
-    t = time()
-    profile_point_dataclass1(init, rest)
-    delta = time() - t
-    print("profile_point_class1: {}".format(delta))
-    t = time()
-    profile_point_dataclass2(init, rest)
-    delta = time() - t
-    print("profile_point_class2: {}".format(delta))
+    timeit(profile_tuple, init, rest)
+    timeit(profile_point0, init, rest)
+    timeit(profile_point1, init, rest)
+    timeit(profile_point_dataclass0, init, rest)
+    timeit(profile_point_dataclass1, init, rest)
+    timeit(profile_point_dataclass2, init, rest)
+    timeit(profile_point_deepcopy, init, rest)