--- a/python/setup.cfg
+++ b/python/setup.cfg
@@ -1,3 +1,5 @@
 [flake8]
 exclude = ENV
 max-line-length = 99
+per-file-ignores =
+    performance/point_update_performance.py:E402