--- a/python/utils/timeit.py
+++ b/python/utils/timeit.py
@@ -1,0 +1,7 @@
+from time import time
+
+def timeit(func, *args, **kwargs):
+    t = time()
+    func(*args, **kwargs)
+    delta = time() - t
+    print("{} took {:.2f} seconds to execute".format(func.__name__, delta))