I updated State/Ball/Flag to be immutable. For training, we will need to be able to keep track of state history, which means if we mutate State, we need to deepcopy() in order to have a record of what our state was. I did performance analysis on deepcopy(), and it is _awful_.

We pay a slight penalty in efficiency for this PR. I timed 1000 iterations of `next_state` and `handle_input`, these are the results:

Current master:
```
next_state took 0.029 seconds to execute
handle_input took 0.002 seconds to execute
```

This PR:
```
next_state took 0.049 seconds to execute
handle_input took 0.017 seconds to execute
```

I think being 2x less efficient, and avoiding a huge deepcopy() penalty later, is worth it. Additionally, this code is easier to reason about.

<!-- Reviewable:start -->
---
This change is [<img src="https://reviewable.io/review_button.svg" height="34" align="absmiddle" alt="Reviewable"/>](https://reviewable.io/reviews/chauncy-crib/tagprobot-ml/21)
<!-- Reviewable:end -->